fix: Extract docsets into the docset directory
The use of QFile::move() to move the docset into its final location only
works if the temporary directory and the docset directory are on the
same filesystem. Rather than recursively copying the docset directory,
just use a temporary directory within the docset directory. This should
make it nearly certain that they reside on the same filesystem.

Close #1